[spritn3][task4]chore-implement SOLID design
