+++ auth-service/Cargo.toml
@@ -13,4 +13,5 @@ serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 uuid = { version = "1.7.0", features = ["v4", "serde"] }
 reqwest = { version = "0.12.4", default-features = false, features = ["json"] }
+async-trait = "0.1.78"
 [dev-dependencies]
