+++ auth-service/src/app_state.rs
@@ -1,9 +1,9 @@
 use std::sync::Arc;
 use tokio::sync::RwLock;
 
-use crate::services::hashmap_user_store::HashmapUserStore;
+use crate::domain::UserStore;
 
-pub type UserStoreType = Arc<RwLock<HashmapUserStore>>;
+pub type UserStoreType = Arc<RwLock<dyn UserStore + Send + Sync>>;
 
 #[derive(Clone)]
 pub struct AppState {
