+++ auth-service/src/domain/data_stores.rs
@@ -1,0 +1,17 @@
+use super::User;
+
+#[derive(Debug, PartialEq)]
+pub enum UserStoreError {
+    UserAlreadyExists,
+    UserNotFound,
+    InvalidCredentials,
+    UnexpectedError,
+}
+
+#[async_trait::async_trait]
+pub trait UserStore {
+    async fn add_user(&mut self, user: User) -> Result<(), UserStoreError>;
+    async fn get_user(&self, email: &str) -> Result<User, UserStoreError>;
+    async fn validate_user(&self, email: &str, password: &str) -> Result<(), UserStoreError> ;
+}
+
