+++ auth-service/src/domain/mod.rs
@@ -1,5 +1,7 @@
 mod user;
 mod error;
+pub mod data_stores;
 
 pub use user::*;
 pub use error::*;
+pub use data_stores::*;
